Restructured status package to allow for running unit tests on macOS or Windows
