+++ pkg/status/registry_windows.go
@@ -3,35 +3,12 @@
 package status
 
 import (
-	"bytes"
-	"fmt"
-	"io/ioutil"
-	"os"
-	"os/exec"
-	"path/filepath"
 	"strings"
 
 	"github.com/1dustindavis/gorilla/pkg/catalog"
-	"github.com/1dustindavis/gorilla/pkg/config"
-	"github.com/1dustindavis/gorilla/pkg/download"
 	"github.com/1dustindavis/gorilla/pkg/gorillalog"
 	version "github.com/hashicorp/go-version"
 	registry "golang.org/x/sys/windows/registry"
-)
-
-// Application Contiains attributes for an installed application
-type Application struct {
-	Key       string
-	Location  string
-	Name      string
-	Source    string
-	Uninstall string
-	Version   string
-}
-
-var (
-	// RegistryItems contains the status of all of the applications in the registry
-	RegistryItems map[string]Application
 )
 
 func stringInSlice(a string, list []string) bool {
@@ -43,14 +20,15 @@ func stringInSlice(a string, list []string) bool {
 	return false
 }
 
-func recoverRegKey() {
-	if r := recover(); r != nil {
-		fmt.Println("Recovered from ", r)
-	}
-}
-
 func getUninstallKeys() map[string]Application {
-	defer recoverRegKey()
+
+	// Recover when the registry lookup fails
+	defer func() {
+		if r := recover(); r != nil {
+			gorillalog.Warn("Recovered from error while accessing the registry: ", r)
+		}
+	}()
+
 	// Get the Uninstall key from HKLM
 	key, err := registry.OpenKey(registry.LOCAL_MACHINE, `Software\Microsoft\Windows\CurrentVersion\Uninstall`, registry.READ)
 	if err != nil {
@@ -99,60 +77,6 @@ func getUninstallKeys() map[string]Application {
 	return installedItems
 }
 
-func checkScript(catalogItem catalog.Item) (actionNeeded bool, checkErr error) {
-
-	// Write InstallCheckScript to disk as a Powershell file
-	tmpScript := filepath.Join(config.CachePath, "tmpCheckScript.ps1")
-	ioutil.WriteFile(tmpScript, []byte(catalogItem.InstallCheckScript), 0755)
-
-	// Build the command to execute the script
-	psCmd := filepath.Join(os.Getenv("WINDIR"), "system32/", "WindowsPowershell", "v1.0", "powershell.exe")
-	psArgs := []string{"-NoProfile", "-NoLogo", "-NonInteractive", "-WindowStyle", "Normal", "-ExecutionPolicy", "Bypass", "-File", tmpScript}
-
-	// Execute the script
-	cmd := exec.Command(psCmd, psArgs...)
-	var stdout, stderr bytes.Buffer
-	cmd.Stdout = &stdout
-	cmd.Stderr = &stderr
-	err := cmd.Run()
-	cmdSuccess := cmd.ProcessState.Success()
-	outStr, errStr := string(stdout.Bytes()), string(stderr.Bytes())
-
-	// Delete the temporary script
-	os.Remove(tmpScript)
-
-	// Log results
-	gorillalog.Debug("Command Error:", err)
-	gorillalog.Debug("stdout:", outStr)
-	gorillalog.Debug("stderr:", errStr)
-
-	// Install if exit 0
-	actionNeeded = cmdSuccess
-
-	return actionNeeded, checkErr
-}
-
-func checkPath(catalogItem catalog.Item) (actionNeeded bool, checkErr error) {
-	path := filepath.Clean(catalogItem.InstallCheckPath)
-	hash := catalogItem.InstallCheckPathHash
-	gorillalog.Debug("Check Path", path)
-
-	// Confirm that path exists
-	// Install if we get an error
-	if _, checkErr := os.Stat(path); checkErr != nil {
-		actionNeeded = true
-	} else {
-		actionNeeded = false
-	}
-
-	// If a hash is configured, verify it matches
-	if !actionNeeded && hash != "" {
-		actionNeeded = download.Verify(path, hash)
-	}
-
-	return actionNeeded, checkErr
-}
-
 // checkRegistry iterates through the local registry and compiles all installed software
 func checkRegistry(catalogItem catalog.Item, installType string) (actionNeeded bool, checkErr error) {
 	// Iterate through the reg keys to compare with the catalog
@@ -198,26 +122,4 @@ func checkRegistry(catalogItem catalog.Item, installType string) (actionNeeded b
 
 	return actionNeeded, checkErr
 
-}
-
-// CheckStatus determines the method for checking status
-func CheckStatus(catalogItem catalog.Item, installType string) (actionNeeded bool, checkErr error) {
-
-	if catalogItem.InstallCheckScript != "" {
-		gorillalog.Info("Checking status via Script:", catalogItem.DisplayName)
-		return checkScript(catalogItem)
-
-	} else if catalogItem.InstallCheckPath != "" {
-		gorillalog.Info("Checking status via Path:", catalogItem.DisplayName)
-		return checkPath(catalogItem)
-
-	}
-
-	// If needed, populate applications status from the registry
-	if len(RegistryItems) == 0 {
-		RegistryItems = getUninstallKeys()
-	}
-
-	gorillalog.Info("Checking status via Registry:", catalogItem.DisplayName)
-	return checkRegistry(catalogItem, installType)
 }
